Fixed the translation for "model provider"
Propose to use "提供商" for the word "provider" throughout the doc instead of "供应商," which is also used for the word "vendor" in other topics of the doc.  
Using different translations for the terms "provider" as in "model provider" and "vendor" as in "tool vendor" helps avoid confusion throughout the document. Please consider.